requirements.txt
torch>=2.1.0
torch-geometric>=2.4.0
torchdiffeq>=0.2.3
numpy>=1.24
scipy>=1.10
pandas>=2.0
scikit-learn>=1.3
sentence-transformers>=2.2.2
transformers>=4.36
faiss-cpu>=1.7.4
networkx>=3.1
redis>=5.0
psycopg2-binary>=2.9
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.7
seaborn>=0.13
datasets>=2.15
huggingface-hub>=0.20
pytest>=7.4
